requirements.txt
flask
openai-whisper
numpy<2
ffmpeg-python